moved the funtions that weren't directly related to our commands to a utility module.
